// frontend/src/lib/enums/component-id.ts
export enum ComponentId {
  Navigation = "navigation",
  Schedule = "schedule",
  Highlight = "highlight",
  LanguageSwitch = "language-switch",
  VideoPromotion = "video-promotion",
  ScrollIndicator = "scroll-indicator",
  BackToTop = "back-to-top",
  ChatBot = "chat-bot",
  Operation = "operation"
}